requirements.txt
alembic
async-timeout
# 3.5.7 works!
gremlinpython>=3.5.8,<3.6
# 3.6.0, 3.7.0, and 3.7.1 do not work!
# gremlinpython>=3.7.1
ipykernel
networkx
numpy
psycopg2-binary
pyjson5
pytest
python-dotenv
requests
SQLAlchemy
tqdm
pgcli
